More bug fixes and cleanup
